Assignment #4:  Finished till Section #8
8.  Implement TabBarViewController behavior
Connect the tab bar buttons to a shared action function. In that
function, add the view of the selected view controller to the
contentView.
It's useful to track what the currently selected button and view
controller is in a property, so you can unselect the button or remove
the view controller's view when another button is selected.
Guide: Creating a Custom Tab Bar